backend/requirements.txt
tensorflow==2.19.0
h5py==3.15.0
numpy==2.0.2
opencv-python-headless==4.12.0.88
Flask>=2.0
pymongo>=4.0
python-dotenv>=1.0
PyJWT>=2.0
flask_mail>=0.9.1
Pillow>=9.0
boto3>=1.26.0
botocore>=1.29.0
requests>=2.28.0
